requirements.txt
pandora-chatgpt[cloud]
pandora-chatgpt[api]
pandora-chatgpt

httpx~=0.23.3
requests~=2.28.2
waitress~=2.1.2
Flask~=2.2.5